+++ CHANGELOG.md
@@ -21,6 +21,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Emit new `SHARD_ALLOCATION_CHANGED` event.
 - Emit new `SHARD_ALLOCATION_NEW` event.
 - Events streaming interface (backed by store).
+- Fine-grained log level configuration.
 - Grafana annotations backend (through JSON datasource).
 - Periodically emit snapshot events.
 - ROADMAP.md to document "nearby" versions and their "aim".